updated files where changes are required while testing

+++ server/controllers/userAuthController.js
@@ -13,8 +13,7 @@ TODO
 const Users = require('../models/userModel'); // User database model.
 const Roles = require('../models/roleModel'); // User roles database model.
 const waitlist= require('../models/signUpExecModel')
-const VolunteerChoiceModel= require('../models/taskInfoModel')
-const AssignedModel= require('../models/taskInfoModel')
+const { VolunteerChoiceModel, AssignedModel } = require('../models/taskInfoModel')
 
 const generateToken = require('../utils/generateToken');
 
@@ -198,6 +197,12 @@ class userAuth{
 				{ new: true }
 			);
 			// user has been updated
+			if (!updatedUser) {
+				return res.status(404).json({
+					success: false,
+					error: "No user found with the provided ID"
+				});
+			}
 			return res.status(200).json({
 				success: true,
 				message: "User updated successfully",
@@ -215,15 +220,20 @@ class userAuth{
 	static deleteUserData = async(req, res) => {
 		try{
 			const user_id = req.user_id;
-			console.log(user_id)  
-			await Promise.all([
+			const del_user = await Users.findOne({_id: user_id })
+			if (!del_user) {
+				return res.status(404).json({
+					success: false,
+					error: "No user found with the provided ID"
+				});
+			}
+			const del = await Promise.all([
 				Roles.deleteMany({ user_id:user_id }),
 				waitlist.deleteMany({user_id: user_id }),
 				VolunteerChoiceModel.deleteMany({ user_id: user_id }),
 				AssignedModel.deleteMany({ user_id :user_id }),
 			]);
-
-			const deletedUser = await Users.findOneAndDelete({ _id: id });
+			const deletedUser = await Users.findOneAndDelete({ _id: user_id });
 			// if the provided ID is not there in the schema
 			// user has been deleted
 			return res.status(200).json({
@@ -232,10 +242,11 @@ class userAuth{
 			});
 		}
 		catch (error) {
-		return res.status(500).json({
-			success: false,
-			message: "Internal server error while updating user profile data",
-		});
+			console.log(error)
+			return res.status(500).json({
+				success: false,
+				message: "Internal server error while updating user profile data",			
+			});
 		}
 	}
